requirements.txt
PyQt6==6.5.0
pymysql==1.1.0
mysql-connector-python==8.3.0
SQLAlchemy==2.0.23
pandas==2.1.4
openpyxl==3.1.2
fpdf2==2.7.8
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1
html2text==2020.1.16
matplotlib==3.8.2
seaborn==0.13.0
